bpo-35134: Create Include/cpython/abstract.h
Move abstract.h code surrounded by "#ifndef Py_LIMITED_API"
to a new Include/cpython/abstract.h header file.
